refactor fills location and directors
