+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
+import 'package:pos_dashboard_v1/features/login/views/login_view.dart';
 import 'package:provider/provider.dart';
 import 'package:pos_dashboard_v1/l10n/app_localizations.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
@@ -44,8 +45,8 @@ class PosSystem extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      // home: const LoginView(),
-      home: const DesktopLayout(),
+      home: const LoginView(),
+      // home: const DesktopLayoutBoady(),
     );
   }
 }
